# contracts/openapi/mezun360.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "BTÜ Mezun360 API",
    "description": "Session authentication and M2B owner profile, privacy and audited manual verification. Production MFA remains pending.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/me/verification-requests": {
      "get": {
        "tags": [
          "verification-controller"
        ],
        "operationId": "getOwnVerification",
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {}
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "CSRF token invalid, CORS rejected or access forbidden.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; sanitized detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "503": {
            "description": "Required dependency unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "verification-controller"
        ],
        "operationId": "submitOwnVerification",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Submitted minimal current education snapshot",
            "headers": {
              "ETag": {
                "schema": {}
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "412": {
            "description": "Exact resource ETag required; stale precondition.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "428": {
            "description": "Exact resource ETag required; stale precondition.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "CSRF token invalid, CORS rejected or access forbidden.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; sanitized detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "503": {
            "description": "Required dependency unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VerificationSubmission": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "confirmAccuracy": {
            "type": "boolean"
          }
        },
        "required": [
          "confirmAccuracy"
        ]
      },
      "VerificationSummary": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ]
          },
          "submitted": {
            "type": "boolean"
          },
          "submittedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reviewedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "rejectionReason": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "rejectionReason",
          "reviewedAt",
          "status",
          "submitted",
          "submittedAt"
        ]
      },
...
      "ApiProblem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "maximum": 599,
            "minimum": 400
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            }
          }
        },
        "required": [
          "code",
          "detail",
          "instance",
          "status",
          "title",
          "traceId",
          "type"
        ]
      },
      "FieldViolation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "field",
          "message"
        ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "mezun360-session",
        "in": "cookie"
      }
    }
  }
}